proj/backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
torch
numpy
scikit-learn
pandas
plotly
folium
streamlit
streamlit-folium
